requirements.txt
ccard==0.1.9
mysql_connector
names==0.3.0
pandas==1.2.4
password_generator==0.1.0
python-dotenv==0.19.2
random_address==1.1.1
randomname==0.1.5
